fix: the card and its data were not deleted correctly

+++ src/main/java/com/alex/d/springbootatm/service/AtmServiceImpl.java
@@ -7,6 +7,7 @@ import com.alex.d.springbootatm.model.TransactionModel;
 import com.alex.d.springbootatm.repository.ATMRepository;
 import com.alex.d.springbootatm.repository.BankCardRepository;
 import com.alex.d.springbootatm.repository.TransactionRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -93,7 +94,7 @@ public class AtmServiceImpl implements ATMService {
     public BankCardModel saveCreatedCardToDB() {
         BankCardModel card = new BankCardModel();
         card.setCardNumber(generateCreditCardNumber());
-        card.setPinNumber(hashPassword(generatePinCode()));
+        card.setPinNumber(hashPinCode(generatePinCode()));
         log.info("Card saved into db {} pin code {}", card.getCardNumber(), card.getPinNumber());
         card.setBalance(generateBalance());
         return bankCardRepository.save(card);
@@ -105,7 +106,7 @@ public class AtmServiceImpl implements ATMService {
         BankCardModel card = saveCreatedCardToDB();
         responseDto.setCardNumber(card.getCardNumber());
         responseDto.setPinCode(generatePinCode());
-        log.info("Card and pincode info {} pin code {}",responseDto.getCardNumber(), responseDto.getPinCode());
+        log.info("Card and pincode info {} pin code {}", responseDto.getCardNumber(), responseDto.getPinCode());
         responseDto.setBalance(card.getBalance());
         return responseDto;
     }
@@ -128,7 +129,7 @@ public class AtmServiceImpl implements ATMService {
     }
 
     @Override
-    public BigDecimal checkBalance(String cardNumber) {
+    public BigDecimal checkBalanceByCardNumber(String cardNumber) {
         Optional<BankCardModel> card = bankCardRepository.findByCardNumber(cardNumber);
         return card.get().getBalance();
     }
@@ -137,8 +138,12 @@ public class AtmServiceImpl implements ATMService {
     @Transactional
     public Optional<BankCardModel> deleteCardByNumber(String cardNumber) {
         Optional<BankCardModel> card = bankCardRepository.findByCardNumber(cardNumber);
-        bankCardRepository.delete(card.get());
-        return card;
+        if (card.isPresent()) {
+            bankCardRepository.delete(card.get());
+            return card;
+        } else {
+            throw new EntityNotFoundException("Card not found " + cardNumber);
+        }
     }
 
     @Override
@@ -150,9 +155,9 @@ public class AtmServiceImpl implements ATMService {
     }
 
     @Override
-    public String hashPassword(String password) {
+    public String hashPinCode(String pinCode) {
         BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
-        return encoder.encode(password);
+        return encoder.encode(pinCode);
     }
 
     @Override
